+++ src/chains/Linea.sol
@@ -11,25 +11,25 @@ import {ISettlerActions} from "../ISettlerActions.sol";
 import {ISignatureTransfer} from "permit2/src/interfaces/ISignatureTransfer.sol";
 import {UnknownForkId} from "../core/SettlerErrors.sol";
 
+import {uniswapV3InitHash, IUniswapV3Callback} from "../core/univ3forks/UniswapV3.sol";
+import {sushiswapV3Factory, sushiswapV3ForkId} from "../core/univ3forks/SushiswapV3.sol";
 import {
-    uniswapV3BlastFactory,
-    uniswapV3InitHash,
-    uniswapV3ForkId,
-    IUniswapV3Callback
-} from "../core/univ3forks/UniswapV3.sol";
-import {thrusterFactory, thrusterInitHash, thrusterForkId} from "../core/univ3forks/Thruster.sol";
+    pancakeSwapV3Factory,
+    pancakeSwapV3InitHash,
+    pancakeSwapV3ForkId,
+    IPancakeSwapV3Callback
+} from "../core/univ3forks/PancakeSwapV3.sol";
 import {IAlgebraCallback} from "../core/univ3forks/Algebra.sol";
-import {bladeSwapFactory, bladeSwapInitHash, bladeSwapForkId} from "../core/univ3forks/BladeSwap.sol";
-import {fenixFactory, fenixInitHash, fenixForkId} from "../core/univ3forks/Fenix.sol";
+import {lynexFactory, lynexInitHash, lynexForkId} from "../core/univ3forks/Lynex.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../SettlerAbstract.sol";
 import {AbstractContext} from "../Context.sol";
 import {Permit2PaymentAbstract} from "../core/Permit2PaymentAbstract.sol";
 
-abstract contract BlastMixin is FreeMemory, SettlerBase {
+abstract contract LineaMixin is FreeMemory, SettlerBase {
     constructor() {
-        assert(block.chainid == 81457 || block.chainid == 31337);
+        assert(block.chainid == 59144 || block.chainid == 31337);
     }
 
     function _dispatch(uint256 i, bytes4 action, bytes calldata data)
@@ -48,21 +48,17 @@ abstract contract BlastMixin is FreeMemory, SettlerBase {
         override
         returns (address factory, bytes32 initHash, uint32 callbackSelector)
     {
-        if (forkId == uniswapV3ForkId) {
-            factory = uniswapV3BlastFactory;
+        if (forkId == pancakeSwapV3ForkId) {
+            factory = pancakeSwapV3Factory;
+            initHash = pancakeSwapV3InitHash;
+            callbackSelector = uint32(IPancakeSwapV3Callback.pancakeV3SwapCallback.selector);
+        } else if (forkId == sushiswapV3ForkId) {
+            factory = sushiswapV3Factory;
             initHash = uniswapV3InitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
-        } else if (forkId == thrusterForkId) {
-            factory = thrusterFactory;
-            initHash = thrusterInitHash;
-            callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
-        } else if (forkId == bladeSwapForkId) {
-            factory = bladeSwapFactory;
-            initHash = bladeSwapInitHash;
-            callbackSelector = uint32(IAlgebraCallback.algebraSwapCallback.selector);
-        } else if (forkId == fenixForkId) {
-            factory = fenixFactory;
-            initHash = fenixInitHash;
+        } else if (forkId == lynexForkId) {
+            factory = lynexFactory;
+            initHash = lynexInitHash;
             callbackSelector = uint32(IAlgebraCallback.algebraSwapCallback.selector);
         } else {
             revert UnknownForkId(forkId);
@@ -71,7 +67,7 @@ abstract contract BlastMixin is FreeMemory, SettlerBase {
 }
 
 /// @custom:security-contact security@0x.org
-contract BlastSettler is Settler, BlastMixin {
+contract LineaSettler is Settler, LineaMixin {
     constructor(bytes20 gitCommit) SettlerBase(gitCommit) {}
 
     function _dispatchVIP(bytes4 action, bytes calldata data) internal override DANGEROUS_freeMemory returns (bool) {
@@ -90,7 +86,7 @@ contract BlastSettler is Settler, BlastMixin {
 
     function _dispatch(uint256 i, bytes4 action, bytes calldata data)
         internal
-        override(SettlerAbstract, SettlerBase, BlastMixin)
+        override(SettlerAbstract, SettlerBase, LineaMixin)
         returns (bool)
     {
         return super._dispatch(i, action, data);
@@ -102,7 +98,7 @@ contract BlastSettler is Settler, BlastMixin {
 }
 
 /// @custom:security-contact security@0x.org
-contract BlastSettlerMetaTxn is SettlerMetaTxn, BlastMixin {
+contract LineaSettlerMetaTxn is SettlerMetaTxn, LineaMixin {
     constructor(bytes20 gitCommit) SettlerBase(gitCommit) {}
 
     function _dispatchVIP(bytes4 action, bytes calldata data, bytes calldata sig)
@@ -117,7 +113,7 @@ contract BlastSettlerMetaTxn is SettlerMetaTxn, BlastMixin {
     // Solidity inheritance is stupid
     function _dispatch(uint256 i, bytes4 action, bytes calldata data)
         internal
-        override(SettlerAbstract, SettlerBase, BlastMixin)
+        override(SettlerAbstract, SettlerBase, LineaMixin)
         returns (bool)
     {
         return super._dispatch(i, action, data);