+++ src/core/univ3forks/Lynex.sol
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.25;
+
+address constant lynexFactory = 0x9A89490F1056A7BC607EC53F93b921fE666A2C48;
+bytes32 constant lynexInitHash = 0xc65e01e65f37c1ec2735556a24a9c10e4c33b2613ad486dd8209d465524bc3f4;
+uint8 constant lynexForkId = 13;