+++ src/core/univ3forks/VelodromeSlipstream.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.25;
 
-address constant velodromeFactory = 0x548118C7E0B865C2CfA94D15EC86B666468ac758;
-bytes32 constant velodromeInitHash = 0x3e17c3f6d9f39d14b65192404b8d70a2f921655d3f7f5e7481ab3fcf0756e8ea;
+address constant velodromeFactory = 0xCc0bDDB707055e04e497aB22a59c2aF4391cd12F;
+bytes32 constant velodromeInitHash = 0x339492e30b7a68609e535da9b0773082bfe60230ca47639ee5566007d525f5a7;
 uint8 constant velodromeForkId = 4;